수정: 단위테스트 중 발견된 수정사항 해결
1. list.tsx 잘못된 colSpan 삭제
2. profileImage.tsx에서 전 프로필 이미지 없으면 삭제 로직 안 거치도록 수정
3. setting.ts getBoardList케이스에서 search 조건 수정
4. index.tsx에서 게시글쪽 디자인 수정

+++ web/pages/api/imageUpload/profileImage.ts
@@ -42,9 +42,11 @@ app.post(uploadImage.single("file"), async (req: any, res: NextApiResponse) => {
       };
 
       const prevProfile = await prisma.members.findUnique(option);
-      const prevProfileImage = prevProfile!.userImage as string;
 
-      bucket.file(prevProfileImage).delete();
+      if (prevProfile!.userImage) {
+        const prevProfileImage = prevProfile!.userImage as string;
+        bucket.file(prevProfileImage).delete();
+      }
 
       const blob = bucket.file(newFileName);
       const blobStream = blob.createWriteStream();
